Add .bin and codegen to gitignore, add Gmodel and beamsolver properties to Shapes class. Update set method for Shapes class.
Add .bin and codegen to gitignore, add Gmodel and beamsolver properties to Shapes class. Update set method for Shapes class.
Add .bin and codegen to gitignore, add Gmodel and beamsolver properties to Shapes class. Update set method for Shapes class. 

This commit adds the .bin and codegen files to the gitignore file. It also adds the Gmodel and beamsolver properties to the Shapes class, along with updating the set method for this class.