Add modeloptions class and update isSelfContact property in shapesoptions
The commit adds a new file called modeloptions.m, which defines a class with properties related to model options. The properties include Color, ColorMap, ColorAxis, Display, LineStyle, isNonlinear, and isAssembled.

In addition to adding the new file, the commit also updates the shapesoptions.m file. Specifically, it changes the value of the isSelfContact property from false to true.