Add gaussQuadrature function for Gaussian quadrature
This commit adds the gaussQuadrature function to the tools/gauss_quad/gaussQuadrature.m file. The function calculates the weights and points for Gaussian quadrature based on the given precision. If no precision is provided, it defaults to 20. The precision must be between 2 and 66, otherwise an assertion error will be thrown.